% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\docType{data}
\name{error_message}
\alias{error_message}
\title{error_messages}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 20 rows and 2 columns.
}
\usage{
error_message
}
\description{
error_messages
}
\keyword{datasets}
